Extract forecast data building
